修改EventRequest資料夾位置、修改EventPhotoEntity名稱
