+++ src/main/java/com/lifespace/repository/EventPhotoRepository.java
@@ -2,8 +2,8 @@ package com.lifespace.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.lifespace.entity.EventPhotoEntity;
+import com.lifespace.entity.EventPhoto;
 
-public interface EventPhotoRepository extends JpaRepository<EventPhotoEntity,String>{
+public interface EventPhotoRepository extends JpaRepository<EventPhoto,String>{
 
 }
