+++ src/main/java/com/lifespace/service/EventPhotoService.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.lifespace.entity.EventPhotoEntity;
+import com.lifespace.entity.EventPhoto;
 import com.lifespace.repository.EventPhotoRepository;
 
 @Service("eventPhotoService")
@@ -15,11 +15,11 @@ public class EventPhotoService {
 	@Autowired
 	EventPhotoRepository repository;
 
-	public void addEventPhoto(EventPhotoEntity eventPhotoEntity) {
+	public void addEventPhoto(EventPhoto eventPhotoEntity) {
 		repository.save(eventPhotoEntity);
 	}
 
-	public void updateEventPhoto(EventPhotoEntity eventPhotoEntity) {
+	public void updateEventPhoto(EventPhoto eventPhotoEntity) {
 		repository.save(eventPhotoEntity);
 	}
 
@@ -30,13 +30,13 @@ public class EventPhotoService {
 //		    repository.deleteById(empno);
 	}
 
-	public EventPhotoEntity getOneEventPhoto(String eventPhotono) {
-		Optional<EventPhotoEntity> optional = repository.findById(eventPhotono);
+	public EventPhoto getOneEventPhoto(String eventPhotono) {
+		Optional<EventPhoto> optional = repository.findById(eventPhotono);
 //		return optional.get();
 		return optional.orElse(null);  // public T orElse(T other) : 如果值存在就回傳其值，否則回傳other的值
 	}
 
-	public List<EventPhotoEntity> getAllEventPhoto() {
+	public List<EventPhoto> getAllEventPhoto() {
 		return repository.findAll();
 	}
 }
